+++ packages/zent/typings/index.d.ts
@@ -2,8 +2,8 @@
 
 /// <reference types="react" />
 
-/// <reference path="./lib/Layout.d.ts" />
-/// <reference path="./lib/Icon.d.ts" />
+/// <reference path="./libs/Layout.d.ts" />
+/// <reference path="./libs/Icon.d.ts" />
 
 declare module 'zent' {
   export * from 'zent/lib/layout'
@@ -89,209 +89,6 @@ declare namespace Zent {
   type IRangePickerProps = IDatePickerProps
 
   class RangePicker extends React.Component<IRangePickerProps, any> { }
-
-
-  interface IRadioProps {
-    value: any
-    className?: string
-    prefix?: string
-  }
-
-  class Radio extends React.Component<IRadioProps, any> { }
-
-  namespace Radio {
-    interface IGroupProps {
-      value: any
-      onChange: React.ChangeEventHandler<HTMLInputElement>
-      isValueEqual?: (value1: any, value2: any) => boolean
-      className?: string
-      prefix?: string
-    }
-
-    class Group extends React.Component<IGroupProps, any> { }
-  }
-
-  interface ISelectTrigger {
-    selectedItems?: Array<any>
-    extraFilter?: boolean
-    open?: boolean
-  }
-
-  interface ISelectProps {
-    data: Array<any>
-    value?: any
-    index?: any
-    disabled?: boolean
-    placeholder?: string
-    searchPlaceholder?: string
-    emptyText?: string
-    trigger?: React.Component<ISelectTrigger, any>
-    optionText?: string
-    optionValue?: string
-    onChange?: (event: { target: { type: any, value: any }, preventDefault: () => void, stopPropagation: () => void }, value: any) => void
-    onDelete?: (date: any) => void
-    filter?: (item: any) => boolean
-    onAsyncFilter?: (keyword: string, callback: (data: any) => void) => void
-    onEmptySelected?: (event: React.SyntheticEvent<HTMLSpanElement>, value: any) => void
-    onOpen?: () => void
-    className?: string
-    prefix?: string
-  }
-
-  class Select extends React.Component<ISelectProps, any> { }
-
-  interface ISliderProps {
-    value: [number, number]
-    onChange?: (value: number) => void
-    range?: boolean
-    min?: number
-    max?: number
-    step?: number
-    withInput?: boolean
-    dots?: boolean
-    marks?: Object
-    disabled?: boolean
-    className?: string
-    prefix?: string
-  }
-
-  class Slider extends React.Component<ISliderProps, any> { }
-
-
-  interface ISwitchProps {
-    checked?: boolean
-    onChange?: (checked: boolean) => void
-    disabled?: boolean
-    checkedText?: string
-    uncheckedText?: string
-    loading?: boolean
-    size?: 'default' | 'small'
-    className?: string
-    prefix?: string
-  }
-
-  class Switch extends React.Component<ISwitchProps, any> { }
-
-
-  interface IStepsProps {
-    type?: 'number' | 'card' | 'breadcrumb'
-    current?: number
-    status?: 'finish' | 'error' | 'wait'
-    className?: string
-    prefix?: string
-  }
-
-  class Steps extends React.Component<IStepsProps, any> { }
-
-  namespace Steps {
-    interface IStepProps {
-      title: React.ReactNode
-      description?: React.ReactNode
-    }
-
-    class Step extends React.Component<IStepProps, any> { }
-  }
-
-  interface ITableColumn {
-    title: string
-    name: string
-    width?: number
-    textAign?: 'left' | 'right' | 'center'
-    isMoney?: boolean
-    bodyRender?: (data: any) => React.ReactNode
-  }
-
-  interface ITableProps {
-    columns: Array<ITableColumn>
-    datasets: Array<Object>
-    rowKey?: string
-    sortBy?: string
-    sortType?: 'desc' | 'asc'
-    onChange?: (conf: any) => void
-    emptyLabel?: string
-    selection?: { selectedRowKeys?: Array<string>, onSelect?: (selectedkeys: string, selectedRows: Array<any>, currentRow: number) => void }
-    loading?: boolean
-    getRowConf?: (data: Object, index: number) => { canSelect: boolean, rowClass: string }
-    expandation?: { isExpanded?: boolean, expandRender?: (data: any) => React.ReactNode }
-    autoStick?: boolean
-    autoScroll?: boolean
-    className?: string
-    prefix?: string
-  }
-
-  class Table extends React.Component<ITableProps, any> { }
-
-  interface ITab {
-    tab: string | number
-    id: string | number
-    disabled?: boolean
-  }
-
-  interface ITabsProps {
-    activeId: string
-    type?: 'normal' | 'card' | 'slider'
-    size?: 'normal' | 'huge'
-    align?: 'left' | 'right' | 'center'
-    onTabChange?: (id: string) => void
-    onTabDel?: (id: string) => void
-    onTabAdd?: () => void
-    candel?: boolean
-    canadd?: boolean
-    tabs?: Array<ITab>
-    className?: string
-    prefix?: string
-  }
-
-  class Tabs extends React.Component<ITabsProps, any> { }
-
-  interface ITreeData {
-    id: number | string
-    title: number | string
-    children?: Array<ITreeData>
-    parendId?: string | number
-    expand?: boolean
-    isLeaf?: boolean
-  }
-
-  interface ITreeOperation {
-    name: string
-    icon?: string | React.ReactElement<any>
-    action: (data: ITreeData) => void
-    shouldRender?: (data: ITreeData) => boolean
-  }
-
-  interface ITreeProps {
-    dataType?: 'tree' | 'plain'
-    data: Array<ITreeData>
-    render?: (data: ITreeData) => React.ReactNode
-    operations?: Array<ITreeOperation>
-    foldable?: boolean
-    checkable?: boolean
-    onCheck?: (data: Array<number | string>) => void
-    defaultCheckedKeys?: Array<number | string>
-    disabledCheckedKeys?: Array<number | string>
-    size?: 'medium' | 'small' | 'large'
-    commonStyle?: React.CSSProperties
-    expandAll?: boolean
-    onExpand?: (data: ITreeData, config: { isExpanded: boolean }) => void
-    autoExpandOnSelect?: boolean
-    onSelect?: (data: ITreeData, target: HTMLSpanElement) => void
-    isRoot?: (data: ITreeData) => boolean
-  }
-
-  class Tree extends React.Component<ITreeProps, any> { }
-
-
-
-  interface IBoundingBox {
-    top: number
-    left: number
-    right: number
-    bottom: number
-    width: number
-    height: number
-  }
-
   interface IPopoverProps {
     position: PositionFunction
     cushion?: number
