+++ packages/zent/typings/libs/Swiper.d.ts
@@ -1,0 +1,19 @@
+/// <reference types="react" />
+
+declare module 'zent/lib/swiper' {
+  interface ISwiperProps {
+    className?: string
+    prefix?: string
+    transitionDuration?: number
+    autoplay?: boolean
+    autoplayInterval?: number
+    dots?: boolean
+    dotsColor?: string
+    dotsSize?: 'normal'|'small'|'large'
+    arrows?: boolean
+    arrowsType?: 'dark'|'light'
+    onChange?: Function
+  }
+
+  export class Swiper extends React.Component<ISwiperProps, any> {}
+}
