+++ packages/zent/typings/libs/Switch.d.ts
@@ -1,0 +1,17 @@
+/// <reference types="react" />
+
+declare module 'zent/lib/switch' {
+  interface ISwitchProps {
+    checked?: boolean
+    onChange?: (checked: boolean) => void
+    disabled?: boolean
+    checkedText?: string
+    uncheckedText?: string
+    loading?: boolean
+    size?: 'default' | 'small'
+    className?: string
+    prefix?: string
+  }
+
+  export class Switch extends React.Component<ISwitchProps, any> { }
+}
